Refactor package structure: rename utils package to llm and update references in prompt.go for improved clarity and organization.
